change port identifiers to use uint16_t

+++ src/apps/peer/mainudpserver.c
@@ -37,6 +37,7 @@
 #include "ns_turn_utils.h"
 #include "udpserver.h"
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -58,7 +59,7 @@ static char Usage[] = "Usage: server [options]\n"
 //////////////////////////////////////////////////
 
 int main(int argc, char **argv) {
-  int port = PEER_DEFAULT_PORT;
+  uint16_t port = PEER_DEFAULT_PORT;
   char **local_addr_list = NULL;
   size_t las = 0;
   int verbose = TURN_VERBOSE_NONE;
